+++ modules/demo-app/scripts/mountApp.client.lua
@@ -1,0 +1,12 @@
+-- _G.__DEV__ = true
+local ReplicatedStorage = game:GetService("ReplicatedStorage")
+local Players = game:GetService("Players")
+
+local DemoApp = require(ReplicatedStorage.Packages.DemoApp)
+
+local PlayerGui = Players.LocalPlayer.PlayerGui
+
+local container = Instance.new("Folder")
+container.Parent = PlayerGui
+
+DemoApp.mountApp(container)