+++ modules/demo-app/src/init.lua
@@ -1,0 +1,19 @@
+--!strict
+local Packages = script.Parent
+local ReactRoblox = require(Packages.ReactRoblox)
+local React = require(Packages.React)
+
+local App = require(script.App)
+
+local function mountApp(container)
+	local root = ReactRoblox.createRoot(container)
+
+	root:render(React.createElement(App))
+
+	return root
+end
+
+return {
+	App = App,
+	mountApp = mountApp,
+}