+++ modules/focus-navigation/src/__tests__/FocusNavigationService.spec.lua
@@ -41,6 +41,60 @@ afterEach(function()
 	GuiService.SelectedCoreObject = nil
 end)
 
+describe("engine interface", function()
+	local mountedTree, service
+	afterEach(function()
+		if mountedTree then
+			mountedTree.root:Destroy()
+			mountedTree = nil :: any
+		end
+		if service then
+			service:teardown()
+			service = nil :: any
+		end
+	end)
+
+	local function getMountedGui(mountTarget)
+		return createGuiObjectTree({
+			root = { "ScreenGui", mountTarget },
+			container = { "Frame", "root", { Size = UDim2.fromScale(0.5, 0.5) } },
+			button = { "ImageButton", "container", { SelectionOrder = 1, Size = UDim2.fromScale(0.5, 0.5) } },
+		})
+	end
+
+	it("should be able to focus the correct input property under CoreGui", function()
+		mountedTree = getMountedGui(CoreGui)
+		service = FocusNavigationService.new(EngineInterface.CoreGui)
+		service:focusGuiObject(mountedTree.button, false)
+		expect(GuiService.SelectedCoreObject).toEqual(mountedTree.button)
+		expect(GuiService.SelectedObject).toBeNil()
+	end)
+
+	it("should be able to focus the correct input property under PlayerGui", function()
+		mountedTree = getMountedGui(PlayerGui)
+		service = FocusNavigationService.new(EngineInterface.PlayerGui)
+		service:focusGuiObject(mountedTree.button, false)
+		expect(GuiService.SelectedObject).toEqual(mountedTree.button)
+		expect(GuiService.SelectedCoreObject).toBeNil()
+	end)
+
+	it("should be able to focus child of non-Selectable CoreGui descendant", function()
+		mountedTree = getMountedGui(CoreGui)
+		service = FocusNavigationService.new(EngineInterface.CoreGui)
+		service:focusGuiObject(mountedTree.container, false)
+		expect(GuiService.SelectedCoreObject).toEqual(mountedTree.button)
+		expect(GuiService.SelectedObject).toBeNil()
+	end)
+
+	it("should be able to focus child of non-Selectable Player descendant", function()
+		mountedTree = getMountedGui(PlayerGui)
+		service = FocusNavigationService.new(EngineInterface.CoreGui)
+		service:focusGuiObject(mountedTree.container, false)
+		expect(GuiService.SelectedObject).toEqual(mountedTree.button)
+		expect(GuiService.SelectedCoreObject).toBeNil()
+	end)
+end)
+
 describe("Basic functionality", function()
 	local tree
 	beforeEach(function()